Cast it.

+++ trunk/src/linad99/jacob3.cpp
@@ -209,8 +209,8 @@ void gradient_structure::jacobcalc(int nvar, const uostream& ofs)
 
       // \todo Need test
       // back up the file one buffer size and read forward
-      off_t offset = sizeof(grad_stack_entry)
-        * gradient_structure::GRAD_STACK1->length;
+      off_t offset = (off_t)(sizeof(grad_stack_entry)
+        * gradient_structure::GRAD_STACK1->length);
       lpos = lseek(gradient_structure::GRAD_STACK1->_GRADFILE_PTR,
         -offset, SEEK_CUR);
